+++ src/main/java/Server/Game/Card/Commands/ActsOnLiveReceptors/HitLiveReceptor.java
@@ -1,11 +1,9 @@
-package Server.Game.Card.Commands;
+package Server.Game.Card.Commands.ActsOnLiveReceptors;
 
+import Server.Game.Card.Commands.CommandName;
 import Server.Game.ModelClasses.LiveReceptor;
 
-import java.util.Arrays;
-
-public class HitLiveReceptor extends ConcreteCommand {
-    private LiveReceptor[] receptors;
+public class HitLiveReceptor extends ActOnLiveReceptors {
     private int attackPoints;
 
     public HitLiveReceptor() {
@@ -14,10 +12,6 @@ public class HitLiveReceptor extends ConcreteCommand {
 
     public void setAttackPoints(int attackPoints) {
         this.attackPoints = attackPoints;
-    }
-
-    public void setReceptors(LiveReceptor[] receptors) {
-        this.receptors = receptors;
     }
 
     @Override
