+++ src/main/java/Server/Game/Card/Commands/ActsOnLiveReceptors/OnCreature/RetreatCreature.java
@@ -1,0 +1,24 @@
+package Server.Game.Card.Commands.ActsOnLiveReceptors.OnCreature;
+
+import Common.Receptors.Creature;
+import Server.Game.Card.Commands.CommandName;
+
+public class RetreatCreature extends ActOnCreature {
+    public RetreatCreature() {
+        super(CommandName.RETREAT_CREATURE);
+    }
+
+    @Override
+    public void execute() {
+        if(receptors != null && receptors[0] != null) {
+            ((Creature)receptors[0]).retreat(((Creature)receptors[0]).getSteps());
+        }
+    }
+
+    @Override
+    public void undo() {
+        if(receptors != null && receptors[0] != null) {
+            ((Creature)receptors[0]).advance();
+        }
+    }
+}
