--- src/main/java/Server/Game/Card/Commands/RetreatCreature.java
@@ -1,25 +1,0 @@
-package Server.Game.Card.Commands;
-
-import Common.Receptors.Creature;
-
-public class RetreatCreature extends ConcreteCommand {
-    private Creature creature;
-
-    public RetreatCreature() {
-        super(CommandName.RETREAT_CREATURE);
-    }
-
-    @Override
-    public void execute() {
-        if(creature != null) {
-            creature.retreat(creature.getSteps());
-        }
-    }
-
-    @Override
-    public void undo() {
-        if(creature != null) {
-            creature.advance();
-        }
-    }
-}
