+++ src/main/java/Server/Game/Utils/GameJsonParser.java
@@ -4,7 +4,7 @@ import Common.Receptors.Player;
 import Server.Game.Card.Card;
 import Server.Game.Card.Commands.CommandName;
 import Server.Game.Card.Commands.ConcreteCommand;
-import Server.Game.Card.Commands.CreateCreature;
+import Server.Game.Card.Commands.ActsOnLiveReceptors.OnCreature.CreateCreature;
 import Server.Game.Game;
 import org.json.JSONException;
 import org.json.JSONObject;
